% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NYCTaxi.R
\name{analyzeFile}
\alias{analyzeFile}
\title{The function to analyze a single pair of files}
\usage{
analyzeFile(idx, path, size_bulk = 500000L, verbose = FALSE,
  verbose_debug = FALSE)
}
\arguments{
\item{idx}{the index of the file, ranging from 1 to 12}

\item{path}{the path of the directory holding the data and fare files}

\item{size_bulk}{the size of each bulk to process, limited by memory size}

\item{verbose}{whether to prompt an information when the processing is done}

\item{verbose_debug}{whether to print the error message when reaching the end of file (or potentially other error message)}
}
\value{
a SummaryNYCTaxi S3 class object
}
\description{
Return a SummaryNYCTaxi S3 class object summarizing one pair of data and fare file, which contains the count of occurence for each value of fare amount less the tolls, and the sufficient statistics for the 2 regression tasks
}

