% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_nyc_data}
\alias{download_nyc_data}
\title{Utility function that download NYC Open Data Open Portal data}
\usage{
download_nyc_data(obj, url, years, n, names, ...)
}
\arguments{
\item{obj}{an etl object}

\item{url}{base url to desired file}

\item{years}{a numeric vector giving the years. The default is the most recent year.}

\item{n}{number of observations}

\item{names}{names of the destination files}

\item{...}{arguments passed to \code{\link[etl]{smart_download}}}
}
\description{
Takes in arguments, sych as obj, year, names, 
as well as the stem of the URL and returns the matched file paths.
}
\examples{
\dontrun{
taxi <- etl("nyctaxi", dir = "~/Desktop/nyctaxi")
base_url <- "https://data.cityofnewyork.us/resource/edp9-qgv4.csv"
download_nyc_data(taxi, base_url, 2015, 100, "2015") 
}

}
