% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R, R/etl_load.R, R/etl_transform.R
\name{etl_extract.etl_nyctaxi}
\alias{etl_extract.etl_nyctaxi}
\alias{etl_load.etl_nyctaxi}
\alias{etl_transform.etl_nyctaxi}
\title{Extract NYC Taxi Trip Data from data from NYC Taxi & Limousine Commission}
\usage{
\method{etl_extract}{etl_nyctaxi}(obj, years = as.numeric(format(Sys.Date(),
  "\%Y")), months = 1:12, type = "yellow", ...)

\method{etl_load}{etl_nyctaxi}(obj, years = as.numeric(format(Sys.Date(),
  "\%Y")), months = 1:12, type = "yellow", ...)

\method{etl_transform}{etl_nyctaxi}(obj, years = as.numeric(format(Sys.Date(),
  "\%Y")), months = 1:12, type = "yellow", ...)
}
\arguments{
\item{obj}{an etl object}

\item{years}{a numeric vector giving the years. The default is the most recent year.}

\item{months}{a numeric vector giving the months. The default is January to December.}

\item{type}{a character variable giving the type of data the user wants to download. 
There are four options: yellow (meaning yellow taxi data), 
green (meaning green taxi data), uber, and lyft. Users can only choose one transportation at a time.
The default is \code{yellow}.}

\item{...}{arguments passed to \code{\link[etl]{smart_download}}}
}
\description{
Extract NYC Taxi Trip Data from data from NYC Taxi & Limousine Commission
}
\details{
extract NYC Yellow taxi trip data from Jan 2009 to the most recent month,
Green taxi trip data from Aug 2013 to the most recent month data from NYC Taxi & Limousine Commission, 
Uber trip data from April to September 2014 and January to June 2015,
Lyft weekly-aggregated data from 2015 to the most recent week.

load NYC Yellow taxi trip data from load directory into a sql database, 
the default is a sqlite database
and/or \code{green}. The default is \code{yellow}.

transform NYC Yellow taxi trip data from raw directory to load directory
and/or \code{green}. The default is \code{yellow}.
}
\examples{


\dontrun{
taxi <- etl("nyctaxi", dir = "~/Desktop/nyctaxi")
taxi \%>\% 
   etl_extract(years = 2014, months = 1:12, type = c("uber")) \%>\% 
   etl_transform(years = 2015, months = 1, type = c("green")) \%>\% 
   etl_load(years = 2015, months = 1, type = c("green")) 
}
}
\seealso{
\code{\link[etl]{etl_extract}}

\code{\link[etl]{etl_load}}

\code{\link[etl]{etl_transform}}
}
