% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_transform.R
\name{etl_transform.etl_nyctaxi}
\alias{etl_transform.etl_nyctaxi}
\title{Extract data from NYC Taxi Trips}
\usage{
\method{etl_transform}{etl_nyctaxi}(obj, ...)
}
\description{
Extract data from NYC Taxi Trips
}
\examples{

if (require(RMySQL)) {
 db <- src_mysql(dbname = "nyctaxi", user = "r-user", password = "mypass")
}

taxi <- etl("nyctaxi", db, dir = "~/dumps/nyctaxi/")
taxi \%>\%
  etl_load()
taxi \%>\%
  
}

