\name{NYCTaxi-package}
\alias{NYCTaxi-package}
\alias{NYCTaxi}
\docType{package}
\title{
\packageTitle{NYCTaxi}
}
\description{
\packageDescription{NYCTaxi}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NYCTaxi}
\packageIndices{NYCTaxi}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{NYCTaxi}

Maintainer: \packageMaintainer{NYCTaxi}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
