% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heap_map.R
\name{plot_heap_map}
\alias{plot_heap_map}
\title{Plot a heatmap of two categorical variables.}
\usage{
plot_heap_map(
  dt,
  cat_vars,
  color_high = "lightslateblue",
  color_low = "gray80",
  sep = " | "
)
}
\arguments{
\item{dt}{A data.table containing the data.}

\item{cat_vars}{A character vector of length 2 specifying the names of the
two categorical variables in \code{dt} to be plotted.}

\item{color_high}{The color to represent high counts in the heatmap.}

\item{color_low}{The color to represent low counts in the heatmap.}

\item{sep}{The separator string to use between the proportion (percentage)
and the count in the text labels on the heatmap tiles. Defaults to " | ".}
}
\value{
A ggplot2 object representing the heatmap.
}
\description{
This function takes a data.table and two categorical variables to create a
heatmap visualizing the counts and proportions of their combinations.
}
