% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NYCTaxi.R
\name{reduceListSummaryNYCTaxi}
\alias{reduceListSummaryNYCTaxi}
\title{The function to reduce a list of SummaryNYCTaxi class object}
\usage{
reduceListSummaryNYCTaxi(list_SummmaryNYCTaxi)
}
\arguments{
\item{list_SummmaryNYCTaxi}{a list of SummaryNYCTaxi class object computed by apply the analyzeFile() function}
}
\value{
a list containing the deciles and 2 sets of linear regression coefficients: total amount less the fee vs trip_time (and surcharge)
}
\description{
Combine the count_occurence hash tables from the list and evaluate the deciles. Sum the sufficient statistics for the 2 regression tasks and solve the linear regression problems
}

