% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_join_geospatial_features.R
\name{required_pkgs.step_join_geospatial_features}
\alias{required_pkgs.step_join_geospatial_features}
\title{Required Packages for \code{step_join_geospatial_features}}
\usage{
\method{required_pkgs}{step_join_geospatial_features}(x, ...)
}
\arguments{
\item{x}{A recipe object.}

\item{...}{Not used.}
}
\value{
A character vector with the names of the required packages.
}
\description{
Identifies the packages needed when using
\code{step_join_geospatial_features()} within a recipe.
This ensures that all dependencies are available when the step
is prepared or baked.
}
