% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxi_zones}
\alias{taxi_zones}
\title{NYC Taxi Zone spatial data}
\format{Spatial shapefiles of NYC Taxi Zones.

\describe{
  \item{OBJECTID}{Object ID}
  \item{Shape_Leng}{The length of the shape corresponding to a TLC taxi zone LocationID.}
  \item{Shape_Area}{The area of the shape corresponding to a TLC taxi zone LocationID.}
  \item{zone}{Zone corresponding to a TLC taxi zone LocationID.}
  \item{LocationID}{TLC taxi zone LocationID.}
  \item{borough}{Borough corresponding to a TLC taxi zone LocationID.}
}}
\source{
https://s3.amazonaws.com/nyc-tlc/misc/taxi_zones.zip
}
\usage{
taxi_zones
}
\description{
NYC Taxi Zone spatial data
}
\keyword{datasets}
