% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateSuffStat}
\alias{updateSuffStat}
\title{Function to update the sufficient statistics of linear regression based on a bulk of data}
\usage{
updateSuffStat(xx_xy, y, x_less_ones)
}
\arguments{
\item{xx_xy}{a p-by-(p+1) matrix, [1:p, 1:p] represents the current value of x^Hx and [p+1, 1:p] represents the current value of x^Hy}

\item{y}{a n-by-1 vector, the bulk of observatons}

\item{x_less_ones}{a n-by-(p-1) matrix, the bulk of predictors excluding 1}
}
\value{
the updated sufficient statistic xx_xy
}
\description{
The sufficient statistics of the linear regression is recorded as a p-by-(p+1) matrix which is the row concatenation of x^Hx and x^Hy where p is the number of predictors (including constant 1).
}

