% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R, R/etl_load.R
\name{etl_extract.etl_nyctaxi}
\alias{etl_extract.etl_nyctaxi}
\alias{etl_load.etl_nyctaxi}
\title{Extract data from NYC Taxi Trips}
\usage{
\method{etl_extract}{etl_nyctaxi}(obj, years = as.numeric(format(Sys.Date(),
  "\%Y")), months = 1:12, types = "yellow", ...)

\method{etl_load}{etl_nyctaxi}(obj, year = "2016", month = 1,
  type = "yellow", ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{types}{a character vector of \code{yellow} and/or \code{green}. The default 
is \code{yellow}.}

\item{...}{arguments passed to methods}

\item{year}{a numeric vector giving the years}

\item{month}{a numeric vector giving the months}
}
\description{
Extract data from NYC Taxi Trips
}
\details{
extract data from NYC taxi trips
}
\examples{

taxi <- etl("nyctaxi", dir = "~/Desktop/nyctaxi/")
\dontrun{
taxi \%>\% 
   etl_extract() \%>\% 
   etl_transform() \%>\% 
   etl_load() 
}
}
\seealso{
\code{\link[etl]{etl_extract}}
}
