% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R
\name{etl_extract.etl_nyctaxi}
\alias{etl_extract.etl_nyctaxi}
\title{Extract data from NYC Taxi Trips}
\usage{
\method{etl_extract}{etl_nyctaxi}(obj, ...)
}
\description{
Extract data from NYC Taxi Trips
}
\examples{

taxi <- etl("nyctaxi", dir = "~/dumps/nyctaxi/")
taxi \%>\%
  etl_extract()
}

