% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R, R/etl_transform.R
\name{etl_extract.etl_nyctaxi}
\alias{etl_extract.etl_nyctaxi}
\alias{etl_load.etl_nyctaxi}
\alias{etl_transform.etl_nyctaxi}
\title{Extract data from NYC Taxi Trips}
\usage{
\method{etl_extract}{etl_nyctaxi}(obj, year = 2016, month = 1,
  type = "yellow", ...)

\method{etl_transform}{etl_nyctaxi}(obj, year = 2016, month = 1,
  type = "yellow", ...)

\method{etl_load}{etl_nyctaxi}(obj, year = "2016", month = 1,
  type = "yellow", ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{...}{arguments passed to methods}
}
\description{
Extract data from NYC Taxi Trips
}
\details{
extract data from NYC taxi trips
}
\examples{

taxi <- etl("nyctaxi", dir = "~/dumps/nyctaxi/")
taxi \%>\% 
   etl_extract() \%>\% 
   etl_transform() \%>\% 
   etl_load() 
}
\seealso{
\code{\link[etl]{etl_extract}}
}

