% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NycTaxi-package.R
\docType{package}
\name{NycTaxi-package}
\alias{NycTaxi}
\alias{NycTaxi-package}
\title{NycTaxi: A Data Analysis Project On NYC Taxi Drivers’ Earnings}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

In this package, you will find the tools needed to reproduce the analysis done in the project to help taxi drivers in NYC implement strategies that can help them increase their earnings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://angelfelizr.github.io/NycTaxi/}
  \item \url{https://github.com/AngelFelizR/NycTaxi/}
  \item Report bugs at \url{https://github.com/AngelFelizR/NycTaxi/issues}
}

}
\author{
\strong{Maintainer}: Angel Feliz \email{angel.esteban.feliz@gmail.com} (Personal Web, <https://angelfeliz.com/>)

}
\keyword{internal}
