% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxi_zones}
\alias{taxi_zones}
\title{NYC Taxi Zone spatial data}
\format{An object of class \code{SpatialPolygonsDataFrame} with 263 rows and 6 columns.}
\usage{
taxi_zones
}
\description{
NYC Taxi Zone spatial data
}
\keyword{datasets}
